// frontend/src/lib/api/model/userMeResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: v2.3.0
 */
import type { UserResponse } from './userResponse';
import type { UserSetting } from './userSetting';

export interface UserMeResponse {
  authenticated: boolean;
  user: UserResponse | null;
  settings: UserSetting[] | null;
}